--- a/curl/bct/transform_decl.go
+++ b/curl/bct/transform_decl.go
@@ -1,0 +1,11 @@
+// +build !gccgo,!appengine
+// +build amd64
+
+package bct
+
+import "github.com/iotaledger/iota.go/curl"
+
+var Indices = curl.Indices
+
+//go:noescape
+func transform(lto, hto, lfrom, hfrom *[curl.StateSize]uint, rounds uint)