--- a/curl/bct/transform_generic.go
+++ b/curl/bct/transform_generic.go
@@ -1,0 +1,11 @@
+// +build !amd64 appengine gccgo
+
+package bct
+
+import (
+	"github.com/iotaledger/iota.go/curl"
+)
+
+func transform(lto, hto, lfrom, hfrom *[curl.StateSize]uint, rounds uint) {
+	transformGeneric(lto, hto, lfrom, hfrom, rounds)
+}