:warning: This PR removes all [mam/v1](https://github.com/iotaledger/iota.go/tree/master/mam/v1) related code. :warning: 

MAM uses undocumented Curl calls and it several contains bugs that would be hard to fix.

# Description of change

This PR improves the Curl performance by optimizing the native go implementation as well as providing amd64 assembly. It removes the cgo dependency.  

On amd64 (assembler) compared against the previous cgo implementation:
```
name                old time/op  new time/op  delta
CurlTransaction-48  2.62ms ± 4%  1.55ms ± 3%  -40.86%
CurlHash-48          152µs ± 2%    49µs ± 2%  -68.00%
```

On 386 (go) against the previous cgo implementation:
```
name                old time/op  new time/op  delta
CurlTransaction-48  2.94ms ± 3%  2.72ms ± 3%   -7.66%
CurlHash-48          172µs ± 2%    85µs ± 2%  -50.44%
```

It also integrates https://github.com/iotaledger/hive.go/blob/master/batchhasher/bct_curl.go with additional optimizations.

On amd64 (assembler) compared against the hive.go implementation:
```
name                old time/op  new time/op  delta
CurlTransaction-48  10.7ms ± 5%  6.13ms ± 5%  -42.71%
```

## Type of change

- Enhancement (a non-breaking change which adds functionality)

## How the change has been tested

The existing ginkgo tests have been extended to test for more cases.

## Change checklist

- [x] My code follows the contribution guidelines for this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] I have added tests using ginkgo that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes