@Wollac 2020-09-09
I am referring to the [mam/v1](https://github.com/iotaledger/iota.go/tree/master/mam/v1) package in this repository. So I was naturally assuming this it was MAMv1. But @kuhlmannmarkus, you are right, it might be MAMv0.